+++ src/grammar/formula.rs
@@ -1,6 +1,5 @@
 use std::cmp::{Ord, Ordering, PartialOrd};
-use std::collections::{BTreeSet, HashSet, VecDeque};
-
+use std::collections::{BTreeSet, HashMap, HashSet, VecDeque};
 use std::fmt;
 use std::ops;
 
@@ -17,9 +16,9 @@ use crate::grammar::{ExpressionIdentifier, Scenario};
 use crate::variants::model::AlleleFreq;
 
 #[derive(Shrinkwrap, Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub(crate) struct IUPAC(u8);
+pub(crate) struct Iupac(u8);
 
-impl IUPAC {
+impl Iupac {
     pub(crate) fn contains(&self, base: u8) -> bool {
         if base == **self {
             return true;
@@ -45,7 +44,7 @@ impl IUPAC {
 #[grammar = "grammar/formula.pest"]
 pub(crate) struct FormulaParser;
 
-#[derive(PartialEq, Eq, Clone, Debug, Hash)]
+#[derive(PartialEq, Eq, Clone, Debug, Hash, PartialOrd, Ord)]
 pub(crate) enum Formula {
     Conjunction { operands: Vec<Formula> },
     Disjunction { operands: Vec<Formula> },
@@ -53,7 +52,36 @@ pub(crate) enum Formula {
     Terminal(FormulaTerminal),
 }
 
-#[derive(PartialEq, Eq, Clone, Debug, Hash)]
+impl From<NormalizedFormula> for Formula {
+    fn from(formula: NormalizedFormula) -> Self {
+        fn from_normalized(formula: NormalizedFormula) -> Formula {
+            match formula {
+                NormalizedFormula::Conjunction { operands } => Formula::Conjunction {
+                    operands: operands.into_iter().map(from_normalized).collect(),
+                },
+                NormalizedFormula::Disjunction { operands } => Formula::Disjunction {
+                    operands: operands.into_iter().map(from_normalized).collect(),
+                },
+                NormalizedFormula::Atom { sample, vafs } => {
+                    Formula::Terminal(FormulaTerminal::Atom { sample, vafs })
+                }
+                NormalizedFormula::Variant {
+                    altbase,
+                    positive,
+                    refbase,
+                } => Formula::Terminal(FormulaTerminal::Variant {
+                    altbase,
+                    positive,
+                    refbase,
+                }),
+                NormalizedFormula::False => Formula::Terminal(FormulaTerminal::False),
+            }
+        }
+        from_normalized(formula)
+    }
+}
+
+#[derive(PartialEq, Eq, Clone, Debug, Hash, PartialOrd, Ord)]
 pub(crate) enum FormulaTerminal {
     Atom {
         sample: String,
@@ -61,8 +89,8 @@ pub(crate) enum FormulaTerminal {
     },
     Variant {
         positive: bool,
-        refbase: IUPAC,
-        altbase: IUPAC,
+        refbase: Iupac,
+        altbase: Iupac,
     },
     Expression {
         identifier: ExpressionIdentifier,
@@ -72,7 +100,7 @@ pub(crate) enum FormulaTerminal {
 }
 
 impl FormulaTerminal {
-    fn merge(&mut self, other: &FormulaTerminal) {
+    fn merge_conjunctions(&mut self, other: &FormulaTerminal) {
         match (self, other) {
             (
                 FormulaTerminal::Atom {
@@ -83,44 +111,67 @@ impl FormulaTerminal {
                     sample: sample_b,
                     vafs: vafs_b,
                 },
-            ) if sample_a == sample_b => {
-                match (&vafs_a, &vafs_b) {
-                    (VAFSpectrum::Range(a), VAFSpectrum::Range(b)) => {
-                        match a.overlap(b) {
-                            VAFRangeOverlap::None => *vafs_a = VAFSpectrum::empty(),
-                            VAFRangeOverlap::Contains | VAFRangeOverlap::Equal => (), // nothing to do
-                            VAFRangeOverlap::Contained => *vafs_a = vafs_b.clone(),
-                            VAFRangeOverlap::Start => {
-                                *vafs_a = VAFSpectrum::Range(VAFRange {
-                                    inner: b.start..a.end,
-                                    left_exclusive: b.left_exclusive,
-                                    right_exclusive: a.right_exclusive,
-                                })
-                            }
-                            VAFRangeOverlap::End => {
-                                *vafs_a = VAFSpectrum::Range(VAFRange {
-                                    inner: a.start..b.end,
-                                    left_exclusive: a.left_exclusive,
-                                    right_exclusive: b.right_exclusive,
-                                })
-                            }
-                        }
-                    }
-                    (VAFSpectrum::Range(a), VAFSpectrum::Set(b)) => {
-                        *vafs_a = VAFSpectrum::Set(
-                            b.iter().filter(|vaf| a.contains(**vaf)).cloned().collect(),
-                        );
-                    }
-                    (VAFSpectrum::Set(a), VAFSpectrum::Range(b)) => {
-                        *vafs_a = VAFSpectrum::Set(
-                            a.iter().filter(|vaf| b.contains(**vaf)).cloned().collect(),
-                        );
-                    }
-                    (VAFSpectrum::Set(a), VAFSpectrum::Set(b)) => {
-                        *vafs_a = VAFSpectrum::Set(a.intersection(b).cloned().collect());
+            ) if sample_a == sample_b => match (&vafs_a, &vafs_b) {
+                (VAFSpectrum::Range(a), VAFSpectrum::Range(b)) => {
+                    *vafs_a = VAFSpectrum::Range(a & b)
+                }
+                (VAFSpectrum::Range(a), VAFSpectrum::Set(b)) => {
+                    *vafs_a = VAFSpectrum::Set(
+                        b.iter().filter(|vaf| a.contains(**vaf)).cloned().collect(),
+                    );
+                }
+                (VAFSpectrum::Set(a), VAFSpectrum::Range(b)) => {
+                    *vafs_a = VAFSpectrum::Set(
+                        a.iter().filter(|vaf| b.contains(**vaf)).cloned().collect(),
+                    );
+                }
+                (VAFSpectrum::Set(a), VAFSpectrum::Set(b)) => {
+                    *vafs_a = VAFSpectrum::Set(a.intersection(b).cloned().collect());
+                }
+            },
+            _ => {
+                panic!("bug: trying to merge FormulaTerminals that are not both atoms and for the same sample")
+            }
+        }
+    }
+
+    /// Try building the union of two instances of VAFSpectrum.
+    /// Returns None in case ranges do not overlap or a set is not fully contained within a range,
+    /// otherwise returns the union.
+    fn try_merge_disjunction(&self, other: &FormulaTerminal) -> Option<VAFSpectrum> {
+        match (self, other) {
+            (
+                FormulaTerminal::Atom {
+                    sample: sample_a,
+                    vafs: vafs_a,
+                },
+                FormulaTerminal::Atom {
+                    sample: sample_b,
+                    vafs: vafs_b,
+                },
+            ) if sample_a == sample_b => match (&vafs_a, &vafs_b) {
+                (VAFSpectrum::Range(a), VAFSpectrum::Range(b)) => match a.overlap(b) {
+                    VAFRangeOverlap::None => None,
+                    _ => Some(VAFSpectrum::Range((a | b).0)),
+                },
+                (VAFSpectrum::Range(a), VAFSpectrum::Set(b)) => {
+                    if b.iter().all(|v| a.contains(*v)) {
+                        Some(VAFSpectrum::Range(a.clone()))
+                    } else {
+                        None
                     }
                 }
-            }
+                (VAFSpectrum::Set(a), VAFSpectrum::Range(b)) => {
+                    if a.iter().all(|v| b.contains(*v)) {
+                        Some(VAFSpectrum::Range(b.clone()))
+                    } else {
+                        None
+                    }
+                }
+                (VAFSpectrum::Set(a), VAFSpectrum::Set(b)) => {
+                    Some(VAFSpectrum::Set(a.union(b).cloned().collect()))
+                }
+            },
             _ => {
                 panic!("bug: trying to merge FormulaTerminals that are not both atoms and for the same sample")
             }
@@ -318,11 +369,7 @@ impl Into<Expr<FormulaTerminal>> for Formula {
 
 impl Formula {
     pub(crate) fn is_terminal(&self) -> bool {
-        if let Formula::Terminal(_) = self {
-            true
-        } else {
-            false
-        }
+        matches!(self, Formula::Terminal(_))
     }
 
     pub(crate) fn to_terminal(&self) -> Option<&FormulaTerminal> {
@@ -414,10 +461,10 @@ impl Formula {
                             formula.clone()
                         }
                     } else {
-                        Err(errors::Error::UndefinedExpression {
+                        return Err(errors::Error::UndefinedExpression {
                             identifier: identifier.to_string(),
-                        })?;
-                        unreachable!();
+                        }
+                        .into());
                     }
                 } else {
                     Formula::Terminal(terminal.clone())
@@ -467,17 +514,20 @@ impl Formula {
     }
 
     fn merge_atoms(&self) -> Self {
+        let group_operands = |operands: &Vec<Formula>| -> HashMap<Option<String>, Vec<Formula>> {
+            operands.iter().cloned().into_group_map_by(|operand| {
+                if let Formula::Terminal(FormulaTerminal::Atom { sample, vafs: _ }) = operand {
+                    Some(sample.to_owned())
+                } else {
+                    // group all non-atoms together
+                    None
+                }
+            })
+        };
         match self {
             Formula::Conjunction { operands } => {
                 // collect statements per sample
-                let mut grouped_operands = operands.iter().cloned().into_group_map_by(|operand| {
-                    if let Formula::Terminal(FormulaTerminal::Atom { sample, vafs: _ }) = operand {
-                        Some(sample.to_owned())
-                    } else {
-                        // group all non-atoms together
-                        None
-                    }
-                });
+                let mut grouped_operands = group_operands(operands);
 
                 // merge atoms of the same sample
                 for (sample, statements) in &mut grouped_operands {
@@ -485,7 +535,7 @@ impl Formula {
                         let mut merged_statement =
                             statements.pop().unwrap().into_terminal().unwrap();
                         for statement in statements.iter() {
-                            merged_statement.merge(statement.to_terminal().unwrap());
+                            merged_statement.merge_conjunctions(statement.to_terminal().unwrap());
                         }
                         *statements = vec![Formula::Terminal(merged_statement)];
                     } else {
@@ -501,9 +551,68 @@ impl Formula {
                         .collect(),
                 }
             }
-            Formula::Disjunction { operands } => Formula::Disjunction {
-                operands: operands.iter().map(|o| o.merge_atoms()).collect(),
-            },
+            Formula::Disjunction { operands } => {
+                // collect statements per sample
+                let mut grouped_operands = group_operands(operands);
+
+                // merge atoms of the same sample
+                for (sample, statements) in &mut grouped_operands {
+                    if let Some(sample) = sample {
+                        // Sort by start position of VAFRange or minimum of VAFSet.
+                        // The idea is to try to keep merging neighbouring ranges/sets, greedily.
+                        statements.sort_unstable_by_key(|stmt| {
+                            if let Formula::Terminal(FormulaTerminal::Atom { sample: _, vafs }) =
+                                stmt
+                            {
+                                match vafs {
+                                    VAFSpectrum::Set(s) => s.iter().min().copied(),
+                                    VAFSpectrum::Range(r) => Some(r.start),
+                                }
+                            } else {
+                                None
+                            }
+                        });
+                        let mut merged_statements = vec![];
+                        // Pick off the first terminal from the list of statements..
+                        let mut current_statement = statements.remove(0).into_terminal().unwrap();
+                        for statement in statements.iter() {
+                            // then look at the (current) next one if the merge was
+                            // successful. Otherwise, `other_statement` will be the last statement
+                            // that could *not* be merged with `current_statement`
+                            let other_statement = statement.to_terminal().unwrap();
+
+                            // Try merging (i.e. unionise!) the two spectra
+                            if let Some(merged) =
+                                current_statement.try_merge_disjunction(other_statement)
+                            {
+                                // if it succeeds, use the merge result as `current_statement`
+                                current_statement = FormulaTerminal::Atom {
+                                    sample: sample.into(),
+                                    vafs: merged,
+                                };
+                            } else {
+                                // if it fails, stash our `current_statement` and replace it with
+                                // the other statement which couldn't be merged with
+                                // `current_statement`.
+                                merged_statements.push(Formula::Terminal(current_statement));
+                                current_statement = other_statement.clone();
+                            }
+                        }
+                        // don't forget to push the last result.
+                        merged_statements.push(Formula::Terminal(current_statement));
+                        *statements = merged_statements;
+                    } else {
+                        continue;
+                    }
+                }
+                Formula::Disjunction {
+                    operands: grouped_operands
+                        .into_iter()
+                        .map(|(_, statements)| statements)
+                        .flatten()
+                        .collect(),
+                }
+            }
             Formula::Negation { operand } => Formula::Negation {
                 operand: Box::new(operand.merge_atoms()),
             },
@@ -545,13 +654,13 @@ impl Formula {
             Formula::Conjunction { operands } => Formula::Disjunction {
                 operands: operands
                     .iter()
-                    .map(|o| Ok(o.negate(scenario, contig)?))
+                    .map(|o| o.negate(scenario, contig))
                     .collect::<Result<Vec<Formula>>>()?,
             },
             Formula::Disjunction { operands } => Formula::Conjunction {
                 operands: operands
                     .iter()
-                    .map(|o| Ok(o.negate(scenario, contig)?))
+                    .map(|o| o.negate(scenario, contig))
                     .collect::<Result<Vec<Formula>>>()?,
             },
             Formula::Negation { operand } => operand.as_ref().clone(),
@@ -588,7 +697,7 @@ impl Formula {
                             match uvafs {
                                 VAFSpectrum::Set(uvafs) => {
                                     let difference: BTreeSet<_> =
-                                        uvafs.difference(&vafs).cloned().collect();
+                                        uvafs.difference(vafs).cloned().collect();
                                     if !difference.is_empty() {
                                         disjunction.push(VAFSpectrum::Set(difference));
                                     }
@@ -693,7 +802,7 @@ impl Formula {
             Formula::Conjunction { operands } => {
                 let operands = operands
                     .iter()
-                    .map(|o| Ok(o.apply_negations(scenario, contig)?))
+                    .map(|o| o.apply_negations(scenario, contig))
                     .collect::<Result<Vec<Formula>>>()?;
 
                 Formula::Conjunction { operands }
@@ -701,7 +810,7 @@ impl Formula {
             Formula::Disjunction { operands } => Formula::Disjunction {
                 operands: operands
                     .iter()
-                    .map(|o| Ok(o.apply_negations(scenario, contig)?))
+                    .map(|o| o.apply_negations(scenario, contig))
                     .collect::<Result<Vec<Formula>>>()?,
             },
             &Formula::Terminal(FormulaTerminal::Variant {
@@ -726,7 +835,7 @@ impl Formula {
     }
 }
 
-#[derive(PartialEq, Eq, Clone, Debug, Hash)]
+#[derive(PartialEq, Eq, Clone, Debug, Hash, PartialOrd, Ord)]
 pub(crate) enum NormalizedFormula {
     Conjunction {
         operands: Vec<NormalizedFormula>,
@@ -740,8 +849,8 @@ pub(crate) enum NormalizedFormula {
     },
     Variant {
         positive: bool,
-        refbase: IUPAC,
-        altbase: IUPAC,
+        refbase: Iupac,
+        altbase: Iupac,
     },
     False,
 }
@@ -759,19 +868,17 @@ impl std::fmt::Display for NormalizedFormula {
             NormalizedFormula::Atom {
                 sample,
                 vafs: VAFSpectrum::Set(vafs),
-            } => {
-                if vafs.len() > 1 {
+            } => match vafs.len() {
+                1 => format!("{}:{}", sample, vafs.iter().next().unwrap()),
+                x if x > 1 => {
                     format!(
                         "{}:{{{}}}",
                         sample,
                         vafs.iter().map(|vaf| format!("{:.1}", vaf)).join(", "),
                     )
-                } else if vafs.len() == 1 {
-                    format!("{}:{}", sample, vafs.iter().next().unwrap())
-                } else {
-                    "false".to_owned()
                 }
-            }
+                _ => "false".to_owned(),
+            },
             NormalizedFormula::Atom {
                 sample,
                 vafs: VAFSpectrum::Range(vafrange),
@@ -807,7 +914,7 @@ impl std::fmt::Display for NormalizedFormula {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
+#[derive(Clone, Debug, PartialEq, Eq, Hash, PartialOrd, Ord)]
 pub(crate) enum VAFSpectrum {
     Set(BTreeSet<AlleleFreq>),
     Range(VAFRange),
@@ -839,6 +946,7 @@ pub(crate) struct VAFRange {
     right_exclusive: bool,
 }
 
+#[derive(Debug, PartialEq, Eq, Copy, Clone, Hash)]
 pub(crate) enum VAFRangeOverlap {
     Contained,
     Contains,
@@ -849,6 +957,14 @@ pub(crate) enum VAFRangeOverlap {
 }
 
 impl VAFRange {
+    pub(crate) fn empty() -> Self {
+        Self {
+            inner: AlleleFreq(0.0)..AlleleFreq(0.0),
+            left_exclusive: true,
+            right_exclusive: true,
+        }
+    }
+
     pub(crate) fn contains(&self, vaf: AlleleFreq) -> bool {
         match (self.left_exclusive, self.right_exclusive) {
             (true, true) => self.start < vaf && self.end > vaf,
@@ -907,7 +1023,12 @@ impl VAFRange {
             (false, true) => range.end < other_range.end,
             (false, false) => range.end < other_range.end,
         };
-        if range.end <= other_range.start || range.start >= other_range.end {
+        if (range.end < other_range.start || range.start > other_range.end)
+            || (range.end <= other_range.start
+                && (range.right_exclusive || other_range.left_exclusive))
+            || (range.start >= other_range.end
+                && (range.left_exclusive || other_range.right_exclusive))
+        {
             VAFRangeOverlap::None
         } else {
             match (start_is_right_of_start, end_is_left_of_end) {
@@ -964,6 +1085,46 @@ impl VAFRange {
         n_obs as f64 * *freq
     }
 }
+
+use auto_ops::impl_op_ex;
+
+impl_op_ex!(&|a: &VAFRange, b: &VAFRange| -> VAFRange {
+    match a.overlap(b) {
+        VAFRangeOverlap::Contained => a.clone(),
+        VAFRangeOverlap::Contains => b.clone(),
+        VAFRangeOverlap::Start => VAFRange {
+            inner: a.inner.start..b.inner.end,
+            left_exclusive: a.left_exclusive,
+            right_exclusive: b.right_exclusive,
+        },
+        VAFRangeOverlap::End => VAFRange {
+            inner: b.inner.start..a.inner.end,
+            left_exclusive: b.left_exclusive,
+            right_exclusive: a.right_exclusive,
+        },
+        VAFRangeOverlap::Equal => a.clone(),
+        VAFRangeOverlap::None => VAFRange::empty(),
+    }
+});
+
+impl_op_ex!(| |a: &VAFRange, b: &VAFRange| -> (VAFRange, Option<VAFRange>) {
+    match a.overlap(b) {
+        VAFRangeOverlap::Contained => (b.clone(), None),
+        VAFRangeOverlap::Contains => (a.clone(), None),
+        VAFRangeOverlap::Start => (VAFRange {
+            inner: b.inner.start..a.inner.end,
+            left_exclusive: b.left_exclusive,
+            right_exclusive: a.right_exclusive,
+        }, None),
+        VAFRangeOverlap::End => (VAFRange {
+            inner: a.inner.start..b.inner.end,
+            left_exclusive: a.left_exclusive,
+            right_exclusive: b.right_exclusive,
+        }, None),
+        VAFRangeOverlap::Equal => (a.clone(), None),
+        VAFRangeOverlap::None => (a.clone(), Some(b.clone()))
+    }
+});
 
 impl ops::Deref for VAFRange {
     type Target = ops::Range<AlleleFreq>;
@@ -1131,8 +1292,8 @@ where
             let refbase = inner.next().unwrap().as_str().as_bytes()[0];
             let altbase = inner.next().unwrap().as_str().as_bytes()[0];
             Formula::Terminal(FormulaTerminal::Variant {
-                refbase: IUPAC(refbase),
-                altbase: IUPAC(altbase),
+                refbase: Iupac(refbase),
+                altbase: Iupac(altbase),
                 positive: true,
             })
         }
@@ -1188,4 +1349,94 @@ where
         Rule::COMMENT => unreachable!(),
         Rule::iupac => unreachable!(),
     })
+}
+
+#[cfg(test)]
+mod test {
+    use crate::grammar::Scenario;
+    use crate::grammar::{Formula, VAFRange};
+    use crate::variants::model::AlleleFreq;
+
+    #[test]
+    fn test_vaf_range_overlap() {
+        let r1 = VAFRange {
+            inner: AlleleFreq(0.0)..AlleleFreq(0.7),
+            left_exclusive: false,
+            right_exclusive: true,
+        };
+        let r2 = VAFRange {
+            inner: AlleleFreq(0.3)..AlleleFreq(1.0),
+            left_exclusive: false,
+            right_exclusive: false,
+        };
+        let expected = VAFRange {
+            inner: AlleleFreq(0.3)..AlleleFreq(0.7),
+            left_exclusive: false,
+            right_exclusive: true,
+        };
+        assert_eq!(expected, r1 & r2);
+    }
+
+    #[test]
+    fn test_range_conjunction() {
+        let scenario: Scenario = serde_yaml::from_str(
+            r#"samples:
+  normal:
+    resolution: 100
+    universe: "[0.0,1.0]"
+events:
+  full: "normal:[0.0,1.0]"
+  part1: "normal:[0.0,0.7]"
+  part2: "normal:[0.3,1.0]"
+  expected: "normal:[0.3,0.7]""#,
+        )
+        .unwrap();
+        let expected = scenario.events["expected"].clone();
+        let full = scenario.events["full"].clone();
+        let part1 = scenario.events["part1"].clone();
+        let part2 = scenario.events["part2"].clone();
+        let conjunction = Formula::Conjunction {
+            operands: vec![part1, part2],
+        }
+        .normalize(&scenario, "all")
+        .unwrap();
+        assert_eq!(conjunction, expected.normalize(&scenario, "all").unwrap());
+        assert_ne!(conjunction, full.normalize(&scenario, "all").unwrap());
+    }
+
+    #[test]
+    fn test_nested_range_disjunction() {
+        let scenario: Scenario = serde_yaml::from_str(
+            r#"samples:
+  normal:
+    resolution: 100
+    universe: "[0.0,1.0]"
+events:
+  full: "(normal:[0.0, 0.25] | normal:[0.5,0.75]) | (normal:[0.25,0.5] | normal:[0.75,1.0]) | normal:[0.1,0.4] | normal:0.1"
+  expected: "normal:[0.0,1.0]""#,
+        )
+            .unwrap();
+        let expected = scenario.events["expected"].clone();
+        let full = scenario.events["full"].clone();
+        let full = full.normalize(&scenario, "all").unwrap();
+        assert_eq!(full, expected.normalize(&scenario, "all").unwrap());
+    }
+
+    #[test]
+    fn test_two_separate_range_disjunctions() {
+        let scenario: Scenario = serde_yaml::from_str(
+            r#"samples:
+  normal:
+    resolution: 100
+    universe: "[0.0,1.0]"
+events:
+  full: "(normal:[0.0, 0.25] | normal:[0.5,0.6]) | ((normal:[0.25,0.5] | normal:[0.7,0.9]) | normal:[0.9,1.0]) | normal:]0.8,0.9[ | normal:0.75"
+  expected: "normal:[0.0,0.6] | normal:[0.7,1.0]""#,
+        )
+            .unwrap();
+        let expected = scenario.events["expected"].clone();
+        let full = scenario.events["full"].clone();
+        let full = full.normalize(&scenario, "all").unwrap();
+        assert_eq!(full, expected.normalize(&scenario, "all").unwrap());
+    }
 }